move DataTableManager widget to  separate file and renamed test file to be tests of widget
