+++ tests/unit/test_data_table_manager.py
@@ -7,87 +7,11 @@ from dyno_viewer.app_types import TableInfo
 from textual.binding import Binding
 from textual.message import Message
 
-
-class DataTableManager(Widget):
-    """
-    handles pagination and displaying of dynamodb query and scan results
-    """
-
-    BINDINGS = {
-        Binding("[", action="page_decrement", description="prev results", show=True),
-        Binding("]", action="page_increment", description="next results", show=True),
-    }
-
-    table_info = reactive(None)
-    data = reactive([])
-    static_cols = reactive(["pk", "sk"])
-    page_index = reactive(-1)
-
-    class PaginateRequest(Message):
-        pass
-
-    def compose(self):
-        yield DataTable()
-
-    def on_mount(self):
-        table = self.query_one(DataTable)
-        table.focus()
-
-    def action_page_decrement(self):
-        if self.page_index > -1:
-            self.page_index -= 1
-
-    def action_page_increment(self):
-        if self.page_index < len(self.data) - 1:
-            self.page_index += 1
-        else:
-            self.post_message(self.PaginateRequest())
-            self.loading = True
-
-    def watch_table_info(self, new_table: TableInfo):
-        if not new_table:
-            return
-        log.info("table_info updated, updating gsi and other key cols for table")
-
-        key_schema = new_table["keySchema"]
-
-        gsi = new_table["gsi"]
-        gsi_cols = [
-            key for gsi in gsi.values() for key in [gsi["primaryKey"], gsi["sortKey"]]
-        ]
-
-        log.info(f"{len(gsi_cols)} gsi cols")
-
-        self.static_cols = [key_schema["primaryKey"], key_schema["sortKey"], *gsi_cols]
-
-        log.info(f"{len(self.static_cols)} total cols")
-
-    def watch_data(self, new_data):
-        if new_data and self.page_index == -1:
-            self.page_index = 0
-
-    def watch_page_index(self, new_page: int):
-        if new_page != -1 and self.data:
-            table = self.query_one(DataTable)
-            table.clear(columns=True)
-            non_static_cols = {
-                col
-                for data_item in self.data[self.page_index]
-                for col in data_item.keys()
-                if col not in self.static_cols
-            }
-            cols = [*self.static_cols, *non_static_cols]
-            for col in cols:
-                table.add_column(col, key=col)
-            rows = [
-                [item.get(col) for col in cols] for item in self.data[self.page_index]
-            ]
-
-            table.add_rows(rows)
+from dyno_viewer.components.table import DataTableManager
 
 
 class DataTableManagerApp(App):
-    data = reactive(None)
+    data = reactive([])
     table_info = reactive(
         {
             "keySchema": {"primaryKey": "pk", "sortKey": "sk"},
@@ -100,15 +24,13 @@ class DataTableManagerApp(App):
     paginate_count = reactive(4)
 
     def run_query(self):
-        self.paginate_count -= 1
-        if self.paginate_count >= 0:
-            return [
-                {"pk": "customer#33333", "sk": "CUSTOMER", "testAttr": "testy"},
-                {"pk": "customer#33334", "sk": "CUSTOMER", "testAttr": "testy"},
-                {"pk": "customer#33335", "sk": "CUSTOMER", "testAttr": "testy"},
-                {"pk": "customer#33336", "sk": "CUSTOMER", "testAttr": "testy"},
-            ]
-        return []
+
+        return [
+            {"pk": "customer#33333", "sk": "CUSTOMER", "testAttr": "testy"},
+            {"pk": "customer#33334", "sk": "CUSTOMER", "testAttr": "testy"},
+            {"pk": "customer#33335", "sk": "CUSTOMER", "testAttr": "testy"},
+            {"pk": "customer#33336", "sk": "CUSTOMER", "testAttr": "testy"},
+        ]
 
     def compose(self):
         yield DataTableManager().data_bind(
